Merge2 (Thanh Phuc Phu)

+++ music/src/main/java/com/boot/music/repositories/DocumentRepo.java
@@ -5,5 +5,6 @@ import com.boot.music.entity.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DocumentRepo extends JpaRepository<Document, Integer> {
-	List<Document> findByStatusID(short statusID);
+
+	List<Document> findByStatusID(int statusID);
 }
